Reprint Server: Canonical Reprint Server build and release entry points.
