+++ composer.json
@@ -33,7 +33,7 @@
         "test": "cd tests && phpunit --testdox --colors",
         "test:coverage": "cd tests && phpunit --coverage-html coverage/ --coverage-text",
         "build:phar": "./bin/build-reprint-phar.sh",
-        "build:server-plugin": "./bin/build-exporter-plugin.sh",
+        "build:server-plugin": "./bin/build-server-plugin.sh",
         "build:exporter-plugin": "@build:server-plugin",
         "analyze": "phpstan analyze --memory-limit=1G",
         "lint:php": "phpcs --standard=phpcs.xml.dist",
