+++ tests/php56-server-smoke.php
@@ -1,7 +1,7 @@
 <?php
 
 if ($argc !== 2 || !is_dir($argv[1])) {
-    fail_php56_exporter_smoke('Usage: php php56-exporter-smoke.php <plugin-directory>');
+    fail_php56_server_smoke('Usage: php php56-server-smoke.php <plugin-directory>');
 }
 
 $plugin_directory = realpath($argv[1]);
@@ -10,35 +10,35 @@ date_default_timezone_set('UTC');
 
 require $plugin_directory . '/lib.php';
 
-if (_site_export_push_is_supported()) {
-    fail_php56_exporter_smoke('Push must remain disabled below PHP 7.2.');
+if (\WordPress\Reprint\Server\Plugin\push_is_supported()) {
+    fail_php56_server_smoke('Push must remain disabled below PHP 7.2.');
 }
 
-$export_runtime = _site_export_load_exporter_runtime();
-if ($export_runtime === null) {
-    fail_php56_exporter_smoke('The packaged exporter runtime could not be loaded.');
+$server_runtime = \WordPress\Reprint\Server\Plugin\load_server_runtime();
+if ($server_runtime === null) {
+    fail_php56_server_smoke('The packaged server runtime could not be loaded.');
 }
-require_once $export_runtime;
+require_once $server_runtime;
 
-if (!class_exists('Site_Export_HTTP_Server')) {
-    fail_php56_exporter_smoke('The packaged HTTP server class could not be loaded.');
+if (!class_exists('WordPress\\Reprint\\Server\\HTTPServer')) {
+    fail_php56_server_smoke('The packaged HTTP server class could not be loaded.');
 }
-if (Site_Export_HTTP_Server::is_push_endpoint('preflight')) {
-    fail_php56_exporter_smoke('The preflight endpoint was classified as push.');
+if (\WordPress\Reprint\Server\HTTPServer::is_push_endpoint('preflight')) {
+    fail_php56_server_smoke('The preflight endpoint was classified as push.');
 }
 
-$temporary_directory = sys_get_temp_dir() . '/reprint-php56-exporter-' . getmypid();
+$temporary_directory = sys_get_temp_dir() . '/reprint-php56-server-' . getmypid();
 if (!mkdir($temporary_directory, 0700) && !is_dir($temporary_directory)) {
-    fail_php56_exporter_smoke('The temporary exporter directory could not be created.');
+    fail_php56_server_smoke('The temporary server directory could not be created.');
 }
 
 try {
     $docroot = $temporary_directory . '/docroot';
     if (!mkdir($docroot, 0700) && !is_dir($docroot)) {
-        fail_php56_exporter_smoke('The push smoke-test document root could not be created.');
+        fail_php56_server_smoke('The push smoke-test document root could not be created.');
     }
     $reprint_directory = $temporary_directory . '/push-state';
-    $server = new Site_Export_HTTP_Server([
+    $server = new \WordPress\Reprint\Server\HTTPServer([
         'push' => [
             'reprint_directory' => $reprint_directory,
             'docroot' => $docroot,
@@ -61,10 +61,10 @@ try {
         || !isset($push_response['detail'])
         || strpos($push_response['detail'], 'Push endpoints require PHP 7.2 or newer') === false
     ) {
-        fail_php56_exporter_smoke('The packaged HTTP server returned the wrong push version response.');
+        fail_php56_server_smoke('The packaged HTTP server returned the wrong push version response.');
     }
     if (file_exists($reprint_directory)) {
-        fail_php56_exporter_smoke('The packaged HTTP server started push work below PHP 7.2.');
+        fail_php56_server_smoke('The packaged HTTP server started push work below PHP 7.2.');
     }
     rmdir($docroot);
     http_response_code(200);
@@ -73,7 +73,7 @@ try {
     $preflight = endpoint_preflight(['directory' => [$temporary_directory]]);
     ob_end_clean();
     if (!isset($preflight['stats']['php']['version'])) {
-        fail_php56_exporter_smoke('Preflight did not report the PHP version.');
+        fail_php56_server_smoke('Preflight did not report the PHP version.');
     }
 
     ob_start();
@@ -82,23 +82,23 @@ try {
     $output_stream->finish();
     $stream_output = ob_get_clean();
     if ($stream_output !== 'php56-stream') {
-        fail_php56_exporter_smoke('PHP 5.6 did not use the identity streaming fallback.');
+        fail_php56_server_smoke('PHP 5.6 did not use the identity streaming fallback.');
     }
 
     if (strlen(WordPress\Reprint\Server\generate_random_bytes(16)) !== 16) {
-        fail_php56_exporter_smoke('The PHP 5.6 random-byte fallback returned the wrong length.');
+        fail_php56_server_smoke('The PHP 5.6 random-byte fallback returned the wrong length.');
     }
 } finally {
     rmdir($temporary_directory);
 }
 
-echo "PHP 5.6 exporter runtime smoke passed.\n";
+echo "PHP 5.6 server runtime smoke passed.\n";
 
 function plugin_dir_path($path) {
     return dirname($path) . '/';
 }
 
-function fail_php56_exporter_smoke($message) {
+function fail_php56_server_smoke($message) {
     fwrite(STDERR, $message . "\n");
     exit(1);
 }
